refactor: DDD 기반 멀티모듈 구조로 프로젝트 분리
- domain, application, infrastructure, presentation 4개 모듈로 분리
- domain 모듈: jakarta.persistence-api만 의존 (Spring 완전 제거)
- BaseEntity @PrePersist/@PreUpdate로 교체 (@EnableJpaAuditing 제거)
- Flyway SQL → infrastructure 모듈, application.yaml → presentation 모듈
- presentation 모듈이 Spring Boot 실행 진입점 역할 담당

Closes #3
