+++ application/build.gradle
@@ -1,0 +1,4 @@
+dependencies {
+	implementation project(':domain')
+	implementation 'org.springframework.boot:spring-boot-starter'
+}
