+++ domain/build.gradle
@@ -1,0 +1,3 @@
+dependencies {
+	implementation 'jakarta.persistence:jakarta.persistence-api'
+}
