+++ domain/src/main/java/com/haeni/carrot/settle/domain/order/OrderStatus.java
@@ -1,0 +1,9 @@
+package com.haeni.carrot.settle.domain.order;
+
+public enum OrderStatus {
+  CREATED,
+  PAID,
+  CONFIRMED,
+  SETTLED,
+  REFUNDED
+}
