+++ domain/src/main/java/com/haeni/carrot/settle/domain/settlement/PayoutStatus.java
@@ -1,0 +1,7 @@
+package com.haeni.carrot.settle.domain.settlement;
+
+public enum PayoutStatus {
+  PENDING,
+  COMPLETED,
+  FAILED
+}
