+++ domain/src/main/java/com/haeni/carrot/settle/domain/settlement/SettlementStatus.java
@@ -1,0 +1,6 @@
+package com.haeni.carrot.settle.domain.settlement;
+
+public enum SettlementStatus {
+  INCOMPLETED,
+  COMPLETED
+}
