+++ presentation/src/test/java/com/haeni/carrot/settle/ApplicationTests.java
@@ -1,0 +1,13 @@
+package com.haeni.carrot.settle;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.boot.test.context.SpringBootTest;
+import org.springframework.context.annotation.Import;
+
+@Import(TestcontainersConfiguration.class)
+@SpringBootTest
+class ApplicationTests {
+
+  @Test
+  void contextLoads() {}
+}
