+++ presentation/src/test/java/com/haeni/carrot/settle/TestApplication.java
@@ -1,0 +1,10 @@
+package com.haeni.carrot.settle;
+
+import org.springframework.boot.SpringApplication;
+
+public class TestApplication {
+
+  public static void main(String[] args) {
+    SpringApplication.from(Application::main).with(TestcontainersConfiguration.class).run(args);
+  }
+}
