+++ settings.gradle
@@ -1,1 +1,2 @@
 rootProject.name = 'carrot.settle'
+include 'domain', 'application', 'infrastructure', 'presentation'
