--- src/main/java/com/haeni/carrot/settle/domain/order/OrderItem.java
@@ -1,73 +1,0 @@
-package com.haeni.carrot.settle.domain.order;
-
-import com.haeni.carrot.settle.domain.product.Product;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.FetchType;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.JoinColumn;
-import jakarta.persistence.ManyToOne;
-import jakarta.persistence.Table;
-import java.math.BigDecimal;
-
-@Entity
-@Table(name = "order_items")
-public class OrderItem {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "order_id", nullable = false)
-  private Order order;
-
-  @ManyToOne(fetch = FetchType.LAZY)
-  @JoinColumn(name = "product_id", nullable = false)
-  private Product product;
-
-  @Column(nullable = false)
-  private int quantity;
-
-  @Column(nullable = false, precision = 19, scale = 2)
-  private BigDecimal unitPrice;
-
-  @Column(nullable = false, precision = 19, scale = 2)
-  private BigDecimal subtotal;
-
-  protected OrderItem() {}
-
-  public OrderItem(Order order, Product product, int quantity, BigDecimal unitPrice) {
-    this.order = order;
-    this.product = product;
-    this.quantity = quantity;
-    this.unitPrice = unitPrice;
-    this.subtotal = unitPrice.multiply(BigDecimal.valueOf(quantity));
-  }
-
-  public Long getId() {
-    return id;
-  }
-
-  public Order getOrder() {
-    return order;
-  }
-
-  public Product getProduct() {
-    return product;
-  }
-
-  public int getQuantity() {
-    return quantity;
-  }
-
-  public BigDecimal getUnitPrice() {
-    return unitPrice;
-  }
-
-  public BigDecimal getSubtotal() {
-    return subtotal;
-  }
-}
