--- src/main/java/com/haeni/carrot/settle/domain/order/OrderStatus.java
@@ -1,9 +1,0 @@
-package com.haeni.carrot.settle.domain.order;
-
-public enum OrderStatus {
-  CREATED,
-  PAID,
-  CONFIRMED,
-  SETTLED,
-  REFUNDED
-}
