--- src/main/java/com/haeni/carrot/settle/domain/seller/Seller.java
@@ -1,54 +1,0 @@
-package com.haeni.carrot.settle.domain.seller;
-
-import com.haeni.carrot.settle.domain.common.BaseEntity;
-import jakarta.persistence.Column;
-import jakarta.persistence.Entity;
-import jakarta.persistence.EnumType;
-import jakarta.persistence.Enumerated;
-import jakarta.persistence.GeneratedValue;
-import jakarta.persistence.GenerationType;
-import jakarta.persistence.Id;
-import jakarta.persistence.Table;
-
-@Entity
-@Table(name = "sellers")
-public class Seller extends BaseEntity {
-
-  @Id
-  @GeneratedValue(strategy = GenerationType.IDENTITY)
-  private Long id;
-
-  @Column(nullable = false)
-  private String name;
-
-  @Column(nullable = false, unique = true)
-  private String email;
-
-  @Enumerated(EnumType.STRING)
-  @Column(nullable = false)
-  private SellerGrade grade;
-
-  protected Seller() {}
-
-  public Seller(String name, String email, SellerGrade grade) {
-    this.name = name;
-    this.email = email;
-    this.grade = grade;
-  }
-
-  public Long getId() {
-    return id;
-  }
-
-  public String getName() {
-    return name;
-  }
-
-  public String getEmail() {
-    return email;
-  }
-
-  public SellerGrade getGrade() {
-    return grade;
-  }
-}
