--- src/main/java/com/haeni/carrot/settle/domain/settlement/PayoutStatus.java
@@ -1,7 +1,0 @@
-package com.haeni.carrot.settle.domain.settlement;
-
-public enum PayoutStatus {
-  PENDING,
-  COMPLETED,
-  FAILED
-}
