--- src/main/java/com/haeni/carrot/settle/domain/settlement/SettlementStatus.java
@@ -1,6 +1,0 @@
-package com.haeni.carrot.settle.domain.settlement;
-
-public enum SettlementStatus {
-  INCOMPLETED,
-  COMPLETED
-}
