--- src/test/java/com/haeni/carrot/settle/TestcontainersConfiguration.java
@@ -1,24 +1,0 @@
-package com.haeni.carrot.settle;
-
-import org.springframework.boot.test.context.TestConfiguration;
-import org.springframework.boot.testcontainers.service.connection.ServiceConnection;
-import org.springframework.context.annotation.Bean;
-import org.testcontainers.containers.GenericContainer;
-import org.testcontainers.postgresql.PostgreSQLContainer;
-import org.testcontainers.utility.DockerImageName;
-
-@TestConfiguration(proxyBeanMethods = false)
-class TestcontainersConfiguration {
-
-  @Bean
-  @ServiceConnection
-  PostgreSQLContainer postgresContainer() {
-    return new PostgreSQLContainer(DockerImageName.parse("postgres:latest"));
-  }
-
-  @Bean
-  @ServiceConnection(name = "redis")
-  GenericContainer<?> redisContainer() {
-    return new GenericContainer<>(DockerImageName.parse("redis:latest")).withExposedPorts(6379);
-  }
-}
